# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: SAFe Tooling API
  version: "1.0.0"
  description: |
    REST API for the SAFe Tooling platform. Manages ARTs, Teams, Program
    Increments, Features, Stories, PI Objectives, Risks, Dependencies, and
    Capacity Plans.

    All entity IDs are UUID strings. Dates are ISO 8601 strings (YYYY-MM-DD).

    Computed fields (wsjf_score, cost_of_delay, available_capacity) are
    returned in responses but must not be sent in request bodies.
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /teams:
    get:
      tags: [teams]
      summary: List teams
      operationId: list_teams
      parameters:
        - name: art_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by ART
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"

    post:
      tags: [teams]
      summary: Create a team
      operationId: create_team
      description: |
        If art_id is supplied the team is linked to that ART and the ART's
        team_ids list is updated atomically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamCreate"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: ART not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
  /teams/{team_id}:
    parameters:
      - $ref: "#/components/parameters/team_id"
    get:
      tags: [teams]
      summary: Get a team
      operationId: get_team
      responses:
        "200":
          description: Team details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [teams]
      summary: Update a team
      operationId: update_team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [teams]
      summary: Delete a team
      operationId: delete_team
      description: Also removes the team from its ART's team_ids list.
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Program Increments
  # ---------------------------------------------------------------------------
...
components:
...
  parameters:
...
    team_id:
      name: team_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "PI 'abc-123' not found"

    # --- Enums ---
...
    Team:
      type: object
      required: [id, name, member_count, velocity_history]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Alpha
        member_count:
          type: integer
          minimum: 1
        art_id:
          type: string
          format: uuid
          nullable: true
        velocity_history:
          type: array
          items:
            type: integer
          example: [42, 38, 40]
    TeamCreate:
      type: object
      required: [name, member_count]
      properties:
        name:
          type: string
        member_count:
          type: integer
          minimum: 1
        art_id:
          type: string
          format: uuid
          nullable: true
        velocity_history:
          type: array
          items:
            type: integer
          default: []
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        member_count:
          type: integer
          minimum: 1
        art_id:
          type: string
          format: uuid
          nullable: true
        velocity_history:
          type: array
          items:
            type: integer

    # --- PI ---
...